requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
mysqlclient==2.2.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
